+++ src/test/java/com/uber/h3core/TestHierarchy.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019, 2022 Uber Technologies, Inc.
+ * Copyright 2019, 2022-2023 Uber Technologies, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,13 @@
 package com.uber.h3core;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import com.google.common.collect.ImmutableList;
 import com.uber.h3core.exceptions.H3Exception;
+import com.uber.h3core.util.LatLng;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
@@ -74,8 +77,8 @@ public class TestHierarchy extends BaseTestH3Core {
 
     // res 0 pentagon has 5 hexagon children and 1 pentagon child at res 1.
     // Total output will be:
-    //   5 * 7 children of res 1 hexagons
-    //   6 children of res 1 pentagon
+    // 5 * 7 children of res 1 hexagons
+    // 6 children of res 1 pentagon
     assertEquals(5 * 7 + 6, pentagonChildren.size());
 
     // Don't crash
@@ -187,5 +190,106 @@ public class TestHierarchy extends BaseTestH3Core {
   @Test(expected = IllegalArgumentException.class)
   public void testH3ToCenterChildOutOfRange() {
     h3.cellToCenterChild("8928308280fffff", 16);
+  }
+
+  @Test
+  public void testCellToChildPos() {
+    assertEquals(0, h3.cellToChildPos(0x88283080ddfffffL, 8));
+    assertEquals(6, h3.cellToChildPos(0x88283080ddfffffL, 7));
+    assertEquals(41, h3.cellToChildPos("88283080ddfffff", 6));
+  }
+
+  @Test(expected = H3Exception.class)
+  public void testCellToChildPosError() {
+    h3.cellToChildPos(0x88283080ddfffffL, 9);
+  }
+
+  @Test(expected = H3Exception.class)
+  public void testCellToChildPosError2() {
+    h3.cellToChildPos("88283080ddfffff", 9);
+  }
+
+  @Test
+  public void childPosToCell() {
+    assertEquals(0x88283080ddfffffL, h3.childPosToCell(0, 0x88283080ddfffffL, 8));
+    assertEquals(
+        0x88283080ddfffffL, h3.childPosToCell(6, h3.cellToParent(0x88283080ddfffffL, 7), 8));
+    assertEquals(
+        "88283080ddfffff", h3.childPosToCell(41, h3.cellToParentAddress("88283080ddfffff", 6), 8));
+  }
+
+  @Test(expected = H3Exception.class)
+  public void testChildPosToCellError() {
+    h3.childPosToCell(-1, 0x88283080ddfffffL, 9);
+  }
+
+  @Test(expected = H3Exception.class)
+  public void testChildPosToCellError2() {
+    h3.childPosToCell(10000, "88283080ddfffff", 9);
+  }
+
+  @Test
+  public void cellToChildPosRoundTrip() {
+    // These are somewhat arbitrary, but cover a few different parts of the globe
+    List<LatLng> testLatLngs =
+        ImmutableList.of(
+            new LatLng(37.81331899988944, -122.409290778685),
+            new LatLng(64.2868041, 8.7824902),
+            new LatLng(5.8815246, 54.3336044),
+            new LatLng(-41.4486737, 143.918175));
+
+    for (LatLng ll : testLatLngs) {
+      for (int res = 0; res < 16; res++) {
+        long child = h3.latLngToCell(ll.lat, ll.lng, res);
+        long parent = h3.cellToParent(child, 0);
+        long pos = h3.cellToChildPos(child, 0);
+        long cell = h3.childPosToCell(pos, parent, res);
+        assertNotEquals("sanity check that pos is not a reference to child", child, pos);
+        assertEquals("round trip produced the same cell", child, cell);
+      }
+    }
+  }
+
+  @Test
+  public void childPosAndChildrenSize() {
+    // one hexagon, one pentagon
+    for (String index : ImmutableList.of("80bffffffffffff", "80a7fffffffffff")) {
+      for (int res = 0; res < 16; res++) {
+        long count = h3.cellToChildrenSize(index, res);
+        assertTrue(
+            "count has the right order of magnitude",
+            Math.pow(6, res) <= count && count <= Math.pow(7, res));
+
+        String child = h3.childPosToCell(count - 1, index, res);
+        long pos = h3.cellToChildPos(child, 0);
+        assertEquals("got expected round trip", count - 1, pos);
+
+        try {
+          h3.childPosToCell(count, index, res);
+          fail("Should have thrown, one more is out of range");
+        } catch (H3Exception e) {
+          // expected
+          assertEquals(2 /* E_DOMAIN */, e.getCode());
+        }
+      }
+    }
+  }
+
+  @Test(expected = H3Exception.class)
+  public void cellToChildrenSizeError() {
+    // Invalid resolution
+    h3.cellToChildrenSize("88283080ddfffff", 5);
+  }
+
+  @Test(expected = H3Exception.class)
+  public void cellToChildrenSizeError2() {
+    // Invalid resolution
+    h3.cellToChildrenSize(0x88283080ddfffffL, -1);
+  }
+
+  @Test(expected = H3Exception.class)
+  public void cellToChildrenSizeError3() {
+    // Invalid index
+    h3.cellToChildrenSize(0xffffffffffffffffL, 9);
   }
 }