+++ CHANGELOG.md
@@ -7,6 +7,14 @@ file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java), and support
 for the Linux x64 and Darwin x64 platforms.
 
 ## Unreleased Changes
+## [4.1.0] - 2023-01-19
+### Added
+- `cellToChildPos` and `childPosToCell` functions. (#121)
+- Made the `cellToChildrenSize` function public. (#121)
+
+### Changed
+- Upgraded the core library to v4.1.0. (#121)
+
 ## [4.0.2] - 2022-09-21
 ### Changed
 - Upgraded the core library to v4.0.1. (#113)